libreadline is a problematic library.  Autoconf adds a block to the generated configure script that checks to see if we should be passing the --as-needed flag to ld(1).  There are portability reasons this code exists (best I can figure), but it presents a problem when we scan for libreadline.
libreadline specifically has unresolved symbols on at least Fedora and RHEL.  Why?  Well, the developer can choose to provide libtermcap, libncurses, or libncursesw (wide-char support) at compile time which all satisfy the curses API dependency that libreadline has.  When the parted configure script runs, it fails on readline because the --as-needed flag causes the linker to not include libncurses even though we already have that on the link line.

My solution is to remove the --as-needed flag when we scan for libreadline and then later check for rl_completion_matches() in libreadline.