Add a test to record the current EXT2 mkpartfs failure for some sizes.
Run this:

    dev=F
    dd if=/dev/null of=$dev bs=1 seek=20M
    ./parted -s $dev mklabel gpt
    ./parted -s $dev mkpartfs primary ext2 0 16796160B

It fails with this diagnostic:

    Error: Attempt to write sectors 32772-32773 outside of partition on /t/F.

But if you choose a size that's one byte smaller, it works:

    ./parted -s $dev mkpartfs primary ext2 0 16796159B

The difference is in how ext2_mkfs_write_meta computes
the number of block groups.  In the former case, it computes
numgroups = 3.  In the latter, numgroups = 2.
The trouble with the first case is that there isn't enough
space for 3 full block groups in a file system of that size.
Hence the eventual attempt to write beyond the initially-
established end-of-file-system mark.
(cherry picked from commit 2c4db612ed7de2a19554a6af617f59ecbd4ff1ca)