History with undo and redo capabilities.
Author: Matt Davis <mattdavis@gmail.com>

Here is a working version of the history with undo/redo capabilities.  The
idea here was based on a talk with Otavio Salvador who mentioned the concepts
from Vanni Brutto.  The idea being to capture all changes and only apply them
all at once.  This protects the user from performing an unwanted change.  What
I did was capture all disk modifications that are committed to disk, and put
them in a list.  The history manager allows the list to be traversed linearly,
so that a change can be undone, viewed in parted (print command), or reapplied
again.  Nothing actually happens to the physical disk until the 'save' command
is issued.

Jim Meyering suggested that the functionality might be useful to libparted
thus most of the functionality has been moved there.  The stdout displays
(printf) are placed in parted.c as the library should not do any output
printing on its own.

Three commands were added:
1) Undo : Undoes a disk modification
2) Redo : Redoes the most recent 'undone' modification
3) Save : Actually commits the list of non-undone modifications to disk

I feel a bit more testing needs to be done, but I am happy with the results
right now.