Compensate for different behavior in t7000 test.
When the t7000 test was first engineered the "n" did not appear in the
output when testing mkpart & mkpartfs.  ATM, for an unknown reason,
the "n" is being echoed in the resulting string causing test failure
when compared with the errI string.  This patch attempts to make both
behaviors (with and without the echoed "n") pass the test.

* tests/t7000-scripting.sh: Make sure we don't have an " n" at the end
of the output string.