tests: correct t1100-busy-label
This test had two problems: FS-manipulation (creates a FAT fs) even
for non-512-byte sectors, and it had a now-unexpected line in the
expected output of its final command.
* tests/t1100-busy-label.sh: Use require_512_byte_sector_size_,
so that we skip this test when in non-512-sector size mode.
Remove the now-bogus line.
Note: this test is not normally run.
In order to run it, you have to run as root and set two envvars, e.g.,
sudo make check DEVICE_TO_ERASE=/dev/sde DEVICE_TO_ERASE_SIZE=16.4MB