tests: resize FAT and HFS file systems
Since it looks like we're stuck with keeping at least FAT and HFS
file-system resizing support in libparted, we have an obligation to
exercise at least some of that functionality.  For now, I've limited
this test to shrinking well-aligned partitions by a well-behaved
amount.  Curtis Gedak helped come up with a usable test.

* tests/t3000-resize-fs.sh: New file.
Use starting partition offset that is on a cylinder boundary,
and increase the partition size to be larger than 256 MiB.
When resizing, choose the new size to be exactly one cylinder
(8MiB) smaller than the original.