doc: mention the "bios_grub" flag
I don't like the idea of documenting this "flag setting" option,
but it's been in the code for long enough now, and in actual use,
that I feel obliged.  In the longer term, we'll probably deprecate
it and instead add an interface to set a partition's type independent
of mkpart's "type" argument.
* doc/parted.texi (set): Mention bios_grub.