ped_disk_clobber: change semantics
Why?  Because if we're going to clobber a partition table,
it is wasteful and error prone to try to interpret as a partition
table whatever existing data there may be.  It may be corrupt, or
a mix of two or more types.  It is far cleaner just to zero out
the key sectors.

The previous mechanism was like this:
for each partition table type, T,
    # See if the device has a partition of type T.
    # If so, run the partition-table-specific code to clear it.
    if t->ops->probe (dev)
      t->ops->clobber (dev)
Instead, simply zero out the first few and last few sectors.
Except for GPT, for which we exempt the first sector (the pMBR).
In order to zero out even the bits DASD uses for its identifying
strings, we have to clobber at least the 8.5KiB at the beginning
of the disk, so round up to 9KiB.
* libparted/disk.c: Include "labels/pt-tools.h".
(find_disk_type): New helper function.
(ped_disk_clobber_exclude):

Reviewed-by: Karel Zak <kzak@redhat.com>