build: make it clear which sector size is in use upon test failure
* Makefile.am (ss-1024 ss-2048 ss-4096): New targets.
Before this, top-level "make check" would lead to four separate
runs through all the regression tests, one for each of the following
simulated sector sizes: 1024 2048 4096, and one more for the native
(usually 512), and it was not easy to tell which sector size was
in use for a failing test.  Now, we still perform exactly the same
tests, but the sector size is now part of each target name.