libparted: reenable use of BLKPG ioctls
This patch effectively reverses commit 1d8f9bec which removed the code
using the new BLKPG ioctls instead of the old BLKRRPART ioctl to update
the in-kernel partition table.  The reason for this is because BLKRRPART
fails if any partition on the disk is in use, but the BLKPG ioctls allow
you to manipulate the other partitions on the disk without requiring a
reboot.  Also BLKRRPART requires that the kernel understand the
partition table on the disk, which may not always be the case.