tests: t9030 wait for partition removal
The removal of the partition may take enough time that the next
mkpart will fail.  Wait for it to be removed.

Remove scsi_debug size and set the end of the partition to 800s so that
it will fit inside the default scsi_debug size of 8M, even when sector
size is 4Kb. This prevents failures on machines with limited memory.

* tests/t9030-align-check.sh: Wait for partition removal and use
a smaller disk.