build: enable subdir-objects Automake option globally
That option is enabled unconditionally in Automake-NG, and enabling
it also for mainline Automake will help to ensure no unexpected
incompatibilities is introduced in a potential port to Automake-NG.

Moreover, if we have a source file 'sub/foo.c', having it compiled
in 'sub/foo.o' rather than in 'foo.o' is undeniably cleaner and more
natural.

* configure.ac (AM_INIT_AUTOMAKE): Add 'subdir-objects'.
* libparted/fs/Makefile.am (AUTOMAKE_OPTIONS): No need to declare
'subdir-objects' explicitly now.

Copyright-paperwork-exempt: yes