libparted: fix gpt end of disk handling
There are two checks for problems with the end of disk.  The first checks
to make sure that the backup gpt is actually at the end of the disk as it
should be.  The second checks to see that the gpt's idea of where the disk
ends is correct.  The handling of the backup gpt location was wrong because
if you chose not to fix the error, then as soon as you made any changes the
backup would be written to the end of the disk anyhow, only the previous
backup would not be zeroed.

This patch fixes the write path to put the backup gpt where the gpt says
it goes, not where we think the disk ends.  This allows you to choose
not to fix the problems, and the backup gpt will be written to the same
place it was before, not the new end of disk.