libparted: allow some common errors to be ignored
Partitions that overlap or extend beyond the end of the disk are common
errors that usually result in people having to use other tools to correct
because parted refuses to operate when it sees them.  Change these errors
to allow you to ignore them and use parted to correct the problem.