Revert "linux-commit: do not unnecessarily open partition device nodes"
This reverts commit 2a6936fab4d4499a4b812dd330d3db50549029e0.  The commit
disabled flushing the block buffer caches on the partition nodes to ensure
cache consistency on 2.6 kernels, supposedly because this was no longer
required.  It appears this was incorrect, and the caches DO still need
flushed, otherwise a new fs written to the partition device does not show
up in the disk device cache, causing parted to still identify the old fs.